Proposed refactor damas -^> batnav
