+++ src/batnav/online/match/Ship.java
@@ -1,4 +1,4 @@
-package damas.online.match;
+package batnav.online.match;
 
 import com.google.common.collect.Lists;
