+++ src/batnav/online/socket/JSONPacket.java
@@ -1,4 +1,4 @@
-package damas.online.socket;
+package batnav.online.socket;
 
 import org.json.JSONObject;
