+++ src/batnav/utils/Logger.java
@@ -1,4 +1,4 @@
-package damas.utils;
+package batnav.utils;
 
 public class Logger
 {
