+++ src/batnav/utils/RestUtils.java
@@ -1,4 +1,4 @@
-package damas.utils;
+package batnav.utils;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
