--- src/damas/online/match/MatchException.java
@@ -1,2 +1,0 @@
-package damas.online.match;public class MatchException {
-}
